+++ mockserver-core/src/main/java/org/mockserver/server/initialize/ExpectationInitializerLoader.java
@@ -8,8 +8,8 @@ import org.mockserver.log.model.LogEntry;
 import org.mockserver.logging.MockServerLogger;
 import org.mockserver.mock.Expectation;
 import org.mockserver.mock.RequestMatchers;
+import org.mockserver.mock.listeners.MockServerMatcherNotifier.Cause;
 import org.mockserver.serialization.ExpectationSerializer;
-import org.mockserver.ui.MockServerMatcherNotifier.Cause;
 
 import java.lang.reflect.Constructor;