+++ mockserver-netty/src/test/java/org/mockserver/dashboard/DashboardWebSocketHandlerTest.java
@@ -11,9 +11,9 @@ import org.mockserver.logging.MockServerLogger;
 import org.mockserver.mock.Expectation;
 import org.mockserver.mock.HttpState;
 import org.mockserver.mock.RequestMatchers;
+import org.mockserver.mock.listeners.MockServerMatcherNotifier;
 import org.mockserver.model.RequestDefinition;
 import org.mockserver.scheduler.Scheduler;
-import org.mockserver.ui.MockServerMatcherNotifier;
 import org.mockserver.uuid.UUIDService;
 
 import java.util.Arrays;